refactor(gravedigger): Modularize Gravedigger random event - Added Gravedigger config option - Refactored Gravedigger code into its own module 	- Added GravediggerHelper, GravediggerOverlay, and GravediggerItemOverlay - Moved Coffin and Grave classes into Gravedigger module
